Agregada la función de salir

+++ ConsoleFile.java
@@ -12,7 +12,7 @@ public class ConsoleFile{
 	public static String[] arrayListToStrings(ArrayList<String> al)
 	{
 		String []out = new String[al.size()];
-		short count = 0;
+		int count = 0;
 		for (String token : al) {
 			out[count++] = token;
 			//System.out.print(token);
@@ -58,7 +58,7 @@ public class ConsoleFile{
 	public static String[] stringToArray(String str){
 		StringTokenizer st = new StringTokenizer(str,";");
 		String[] out = new String[st.countTokens()];
-		byte count = 0;
+		int count = 0;
 		while(st.hasMoreTokens()){
 			out[count++] = st.nextToken();
 			//System.out.print(token);
